tests: migrate shell command approval elicitation test from elicitation.rs to codex_tool.rs
Moved the test for shell command approval elicitation to codex_tool.rs and removed the now redundant elicitation.rs file. This consolidates related tests and reduces duplication.
